Fixed a small problem with the camera zoom.

+++ Assets/Scripts/scr_Camera.cs
@@ -58,10 +58,10 @@ public class scr_Camera : MonoBehaviour
             Vector2 mousePosition = Camera.main.ScreenToViewportPoint(Input.mousePosition);
             //Vector2 mov = (Input.mousePosition - transform.position) * zoomMoveFactor;
             mousePosition = new Vector2(mousePosition.x - 0.5f, mousePosition.y - 0.5f) * 2 * zoomMoveFactor;
-	    int x = mousePosition.x;
+	    int x = (int)mousePosition.x;
 	    x /= 5;
 	    x *= 5;
-	    int y = mousePosition.x;
+	    int y = (int)mousePosition.x;
 	    y /= 5;
 	    y *= 5;
             moveTowards = new Vector3(moveTowards.x + mousePosition.x, moveTowards.y + mousePosition.y, moveTowards.z + zoomFactor);
